initial integration of sis-filter template tag
