+++ student-info-system/professor/templates/professor/section.html
@@ -95,7 +95,4 @@
     <input type="submit" value="Back" class="btn btn-info"/>
   </form>
 
-  <script>
-    addClickHandlers("{% url 'professor:student' studentid=1234599 %}", "sched_sel");
-  </script>
 {% endblock content %}
