+++ student-info-system/professor/templates/professor/sections.html
@@ -1,15 +1,12 @@
 
 {% extends 'sis/subpage_base.html' %}
 {% load static %}
-{% load render_table from django_tables2 %}
+{% load sis-filter %}
 
 {% block content %}
-<div class="data-table-alt scrollify-me">
-  {% render_table sections 'django_tables2/bootstrap.html' %}
-</div>
-  <script>
+{% sis_filter 'sections' %}
+    <script>
     const handler = insertScrollableDivs(applyHandlers=true);
     handler();
-    addClickHandlers("{% url 'professor:section' sectionid=1234599 %}", "section_row");
   </script>
 {% endblock content %}
