+++ student-info-system/schooladmin/templates/schooladmin/courses.html
@@ -1,26 +1,13 @@
 {% extends 'sis/subpage_base.html' %}
 {% load static %}
-{% load render_table from django_tables2 %}
+{% load sis-filter %}
 {% load bootstrap3 %}
 
 {% block content %}
 <div>
 <button class="btn btn-primary" onclick="location.href='{% url 'schooladmin:course_new' %}';">New Course</button>
 <div>
-<div id="Courses">
-        {% if courses_filter %}
-            <form id="courses_filter" action="" method="get" class="form form-inline" autocomplete="off">
-                {% bootstrap_form courses_filter.form layout='inline' %}
-                <button class="btn btn-primary">Filter</button>
-                {% if courses_has_filter %}
-                    <a  class="btn btn-success" href="{%  url 'schooladmin:courses' %}">Clear</a>
-                {% endif %}
-            </form>
-        {% endif %}
-        <div class="data-table-alt scrollify-me">
-            {% render_table courses_table 'django_tables2/bootstrap.html' %}
-        </div>
-</div>
+{% sis_filter 'courses' %}
 
 </div>
 </div>
@@ -32,6 +19,5 @@ const handler = insertScrollableDivs(applyHandlers=true);
 handler();
 stylizeRange({field:'catalog_number',name:'Cat Num',prefix:'courses'});
 stylizeRange({field:'credits_earned',name:'Cred',prefix:'courses'});
-addClickHandlers("{% url 'schooladmin:course' courseid=1234599 %}", "course_row");
 </script>
 {% endblock %}
