+++ student-info-system/schooladmin/templates/schooladmin/demographics.html
@@ -1,6 +1,5 @@
 {% extends 'sis/subpage_base.html' %}
 {% load static %}
-{% load render_table from django_tables2 %}
 {% load bootstrap3 %}
 {% block content %}
 <h3>Demographic Summary</h3>
