+++ student-info-system/schooladmin/templates/schooladmin/major.html
@@ -1,7 +1,7 @@
 {% extends 'sis/subpage_base.html' %}
 {% load static %}
 {% load bootstrap3 %}
-{% load render_table from django_tables2 %}
+{% load sis-filter %}
 
 {% block content %}
 <dl class="dl-horizontal">
@@ -16,69 +16,17 @@
 </dl>
 
          <h3>Professors:</h3>
-    <div id="Professors" class="data-table ">
-        {% if profs_filter %}
-            <form id="profs_filter" action="" method="get" class="form form-inline" autocomplete="off">
-                {% bootstrap_form profs_filter.form layout='inline' %}
-                <button class="btn btn-primary">Filter</button>
-                {% if profs_has_filter %}
-                    <a  class="btn btn-success" href="{%  url 'schooladmin:major' majorid=major.id %}">Clear</a>
-                {% endif %}
-            </form>
-        {% endif %}
-        <div class="data-table-alt">
-            {% render_table profs_table 'django_tables2/bootstrap.html' %}
-        </div>
-    </div>
+    {% sis_filter 'profs' %}
 
         <h3>Required Courses:</h3>
-    <div id="required">
-        {% if required_filter %}
-            <form id="required_filter" action="" method="get" class="form form-inline" autocomplete="off">
-                {% bootstrap_form required_filter.form layout='inline' %}
-                <button class="btn btn-primary">Filter</button>
-                {% if required_has_filter %}
-                    <a  class="btn btn-success" href="{%  url 'schooladmin:major' majorid=major.id %}">Clear</a>
-                {% endif %}
-            </form>
-        {% endif %}
-        <div class="data-table-alt">
-            {% render_table required_table 'django_tables2/bootstrap.html' %}
-        </div>
-    </div>
+    {% sis_filter 'required' %}
 
         <h3>Courses offered by Major:</h3>
-    <div id="offered">
-        {% if offered_filter %}
-            <form id="offered_filter" action="" method="get" class="form form-inline" autocomplete="off">
-                {% bootstrap_form offered_filter.form layout='inline' %}
-                <button class="btn btn-primary">Filter</button>
-                {% if offered_has_filter %}
-                    <a  class="btn btn-success" href="{%  url 'schooladmin:major' majorid=major.id %}">Clear</a>
-                {% endif %}
-            </form>
-        {% endif %}
-        <div class="data-table-alt">
-            {% render_table offered_table 'django_tables2/bootstrap.html' %}
-        </div>
-    </div>
+    {% sis_filter 'offered' %}
 
     {%  if students_table %}
         <h3>Students with Major:</h3>
-        <div id="students">
-            {% if students_filter %}
-                <form id="students_filter" action="" method="get" class="form form-inline" autocomplete="off">
-                    {% bootstrap_form students_filter.form layout='inline' %}
-                    <button class="btn btn-primary">Filter</button>
-                    {% if students_has_filter %}
-                        <a  class="btn btn-success" href="{%  url 'schooladmin:major' majorid=major.id %}">Clear</a>
-                    {% endif %}
-                </form>
-            {% endif %}
-            <div class="data-table-alt">
-                {% render_table students_table 'django_tables2/bootstrap.html' %}
-            </div>
-        </div>
+        {% sis_filter 'students' %}
     {% endif %}
     <div>
     <input
@@ -97,7 +45,5 @@
     </div>
 <script>
 {% if permit_edit %}addButtonClickHandler('edit_major', '{% url 'schooladmin:major_edit' major.id %}');{% endif %}
-addClickHandlers("{% url 'student:user' userid=1234599 %}", "user_row");
-addClickHandlers("{% url 'student:course' courseid=1234599 %}", "course_row");
 </script>
 {% endblock %}
