+++ student-info-system/schooladmin/templates/schooladmin/majors.html
@@ -1,29 +1,15 @@
 {% extends 'sis/subpage_base.html' %}
 {% load static %}
-{% load render_table from django_tables2 %}
+{% load sis-filter %}
 {% load bootstrap3 %}
 
 {% block content %}
 <div>
 <button class="btn btn-primary" onclick="location.href='{% url 'schooladmin:major_new' %}';">New Major</button>
-    <div id="Majors">
-        {% if majors_filter %}
-            <form id="majors_filter" action="" method="get" class="form form-inline" autocomplete="off">
-                {% bootstrap_form majors_filter.form layout='inline' %}
-                <button class="btn btn-primary">Filter</button>
-                {% if majors_has_filter %}
-                    <a  class="btn btn-success" href="{%  url 'schooladmin:majors' %}">Clear</a>
-                {% endif %}
-            </form>
-        {% endif %}
-        <div class="data-table-alt scrollify-me">
-            {% render_table majors_table 'django_tables2/bootstrap.html' %}
-        </div>
-    </div>
+{% sis_filter 'majors' %}
 </div>
 <script>
 const handler = insertScrollableDivs(applyHandlers=true);
 handler();
-addClickHandlers("{% url 'schooladmin:major' majorid=1234599 %}", "major_row");
 </script>
 {% endblock content %}
