+++ student-info-system/schooladmin/templates/schooladmin/messages.html
@@ -2,6 +2,7 @@
 {% load static %}
 {% load bootstrap3 %}
 {% load render_table from django_tables2 %}
+{% load sis-filter %}
 {% load user_demographics %}
 
 {% block content %}
