+++ student-info-system/schooladmin/templates/schooladmin/professor.html
@@ -1,7 +1,7 @@
 {% extends 'sis/subpage_base.html' %}
 {% load static %}
 {% load bootstrap3 %}
-{% load render_table from django_tables2 %}
+{% load sis-filter %}
 {%  load user_demographics %}
 
 {% block content %}
@@ -23,84 +23,19 @@
     {%  demographic_details user.profile %}
 
         <h3>Received Messages:</h3>
-    <div id="received">
-        {% if received_filter %}
-            <form id="received_filter" action="" method="get" class="form form-inline" autocomplete="off">
-                {% bootstrap_form received_filter.form layout='inline' %}
-                <button class="btn btn-primary">Filter</button>
-                {% if received_has_filter %}
-                    <a  class="btn btn-success" href="{%  url 'schooladmin:user' userid=user.id %}">Clear</a>
-                {% endif %}
-            </form>
-        {% endif %}
-        <div class="data-table-alt">
-            {% render_table received_table 'django_tables2/bootstrap.html' %}
-        </div>
-    </div>
+    {% sis_filter 'received' %}
 
         <h3>Sent Messages:</h3>
-    <div id="sent">
-        {% if sent_filter %}
-            <form id="sent_filter" action="" method="get" class="form form-inline" autocomplete="off">
-                {% bootstrap_form sent_filter.form layout='inline' %}
-                <button class="btn btn-primary">Filter</button>
-                {% if sent_has_filter %}
-                    <a  class="btn btn-success" href="{%  url 'schooladmin:user' userid=user.id %}">Clear</a>
-                {% endif %}
-            </form>
-        {% endif %}
-        <div class="data-table-alt">
-            {% render_table sent_table 'django_tables2/bootstrap.html' %}
-        </div>
-    </div>
+    {% sis_filter 'sent' %}
 
     <h3>Semesters Taught</h3>
-    <div id="semesterstaught">
-        {% if semesters_filter %}
-            <form id="semesters_filter" action="" method="get" class="form form-inline" autocomplete="off">
-                {% bootstrap_form semesters_filter.form layout='inline' %}
-                <button class="btn btn-primary">Filter</button>
-                {% if semesters_has_filter %}
-                    <a  class="btn btn-success" href="{%  url 'schooladmin:user' userid=user.id %}">Clear</a>
-                {% endif %}
-            </form>
-        {% endif %}
-        <div class="data-table-alt">
-            {% render_table semesters_table 'django_tables2/bootstrap.html' %}
-        </div>
-    </div>
+    {% sis_filter 'semesters' %}
 
     <h3>Sections Taught</h3>
-    <div id="sectionstaught">
-        {% if sections_filter %}
-            <form id="sections_filter" action="" method="get" class="form form-inline" autocomplete="off">
-                {% bootstrap_form sections_filter.form layout='inline' %}
-                <button class="btn btn-primary">Filter</button>
-                {% if sections_has_filter %}
-                    <a  class="btn btn-success" href="{%  url 'schooladmin:user' userid=user.id %}">Clear</a>
-                {% endif %}
-            </form>
-        {% endif %}
-        <div class="data-table-alt">
-            {% render_table sections_table 'django_tables2/bootstrap.html' %}
-        </div>
-    </div>
+    {% sis_filter 'sections' %}
 
     <h3>Reference Items for Courses</h3>
-    <div id="items">
-        {% if items_filter %}
-            <form id="items_filter" action="" method="get" class="form form-inline" autocomplete="off">
-                {% bootstrap_form items_filter.form layout='inline' %}
-                <button class="btn btn-primary">Filter</button>
-                {% if items_has_filter %}
-                    <a  class="btn btn-success" href="{%  url 'schooladmin:user' userid=user.id %}">Clear</a>
-                {% endif %}
-            </form>
-        {% endif %}
-        <div class="data-table-alt">
-            {% render_table items_table 'django_tables2/bootstrap.html' %}
-        </div>
-    </div>
+    {% sis_filter 'items' %}
 
   <form method="post">
     {% csrf_token %}
@@ -168,8 +103,5 @@ addButtonClickHandler('new_item', "{% url 'schooladmin:professor_item_new' useri
 addButtonClickHandler('change_password', "{% url 'schooladmin:change_password' user.id %}");
 addButtonClickHandler('edit_professor', "{% url 'schooladmin:user_edit' user.id %}");
 addButtonClickHandler('showitems', "{% url 'schooladmin:professor_items' user.id %}");
-addClickHandlers("{% url 'schooladmin:semester' semester_id=1234599 %}", "semester_row");
-addClickHandlers("{% url 'schooladmin:section' sectionid=1234599 %}", "section_row");
-addClickHandlers("{% url 'schooladmin:professor_item' userid=user.id item_id=1234599 %}", "refitem_row");
 </script>
 {% endblock %}
