+++ student-info-system/schooladmin/templates/schooladmin/professor_items.html
@@ -1,7 +1,7 @@
 {% extends 'sis/subpage_base.html' %}
 {% load static %}
 {% load bootstrap3 %}
-{% load render_table from django_tables2 %}
+{% load sis-filter %}
 
 {% block content %}
 <dl class="dl-horizontal">
@@ -12,20 +12,7 @@
 </dl>
 
     <h3>Reference Items for Courses</h3>
-    <div id="items">
-        {% if items_filter %}
-            <form id="items_filter" action="" method="get" class="form form-inline" autocomplete="off">
-                {% bootstrap_form items_filter.form layout='inline' %}
-                <button class="btn btn-primary">Filter</button>
-                {% if items_has_filter %}
-                    <a  class="btn btn-success" href="{%  url 'schooladmin:user' userid=user.id %}">Clear</a>
-                {% endif %}
-            </form>
-        {% endif %}
-        <div class="data-table-alt">
-            {% render_table items_table 'django_tables2/bootstrap.html' %}
-        </div>
-    </div>
+    {% sis_filter 'items' %}
 
   <form method="post">
     {% csrf_token %}
@@ -45,6 +32,5 @@
   </form>
 <script>
 addButtonClickHandler('new_item', "{% url 'schooladmin:professor_item_new' userid=user.id %}");
-addClickHandlers("{% url 'schooladmin:professor_item' userid=user.id item_id=1234599 %}", "refitem_row");
 </script>
 {% endblock %}
