+++ student-info-system/schooladmin/templates/schooladmin/profile.html
@@ -1,7 +1,7 @@
 {% extends 'sis/subpage_base.html' %}
 {% load static %}
 {% load bootstrap3 %}
-{% load render_table from django_tables2 %}
+{% load sis-filter %}
 {% load user_demographics %}
 
 {% block content %}
@@ -19,36 +19,10 @@
     {%  demographic_details user.profile %}
 
         <h3>Received Messages:</h3>
-    <div id="received">
-        {% if received_filter %}
-            <form id="received_filter" action="" method="get" class="form form-inline" autocomplete="off">
-                {% bootstrap_form received_filter.form layout='inline' %}
-                <button class="btn btn-primary">Filter</button>
-                {% if received_has_filter %}
-                    <a  class="btn btn-success" href="{%  url 'schooladmin:user' userid=user.id %}">Clear</a>
-                {% endif %}
-            </form>
-        {% endif %}
-        <div class="data-table-alt">
-            {% render_table received_table 'django_tables2/bootstrap.html' %}
-        </div>
-    </div>
+    {% sis_filter 'received' %}
 
         <h3>Sent Messages:</h3>
-    <div id="sent">
-        {% if sent_filter %}
-            <form id="sent_filter" action="" method="get" class="form form-inline" autocomplete="off">
-                {% bootstrap_form sent_filter.form layout='inline' %}
-                <button class="btn btn-primary">Filter</button>
-                {% if sent_has_filter %}
-                    <a  class="btn btn-success" href="{%  url 'schooladmin:user' userid=user.id %}">Clear</a>
-                {% endif %}
-            </form>
-        {% endif %}
-        <div class="data-table-alt">
-            {% render_table sent_table 'django_tables2/bootstrap.html' %}
-        </div>
-    </div>
+    {% sis_filter 'sent' %}
 
   <form method="post">
     {% csrf_token %}
