+++ student-info-system/schooladmin/templates/schooladmin/section.html
@@ -1,7 +1,7 @@
 {% extends 'sis/subpage_base.html' %}
 {% load static %}
 {% load bootstrap3 %}
-{% load render_table from django_tables2 %}
+{% load sis-filter %}
 
 {% block content %}
 <dl class="dl-horizontal">
@@ -26,37 +26,11 @@
   <dd>{{ section.status }}</dd>
 </dl>
 
-        <h3>Reference Items:</h3>
-    <div id="items">
-        {% if secitem_filter %}
-            <form id="secitem_filter" action="" method="get" class="form form-inline" autocomplete="off">
-                {% bootstrap_form secitem_filter.form layout='inline' %}
-                <button class="btn btn-primary">Filter</button>
-                {% if secitem_has_filter %}
-                    <a  class="btn btn-success" href="{%  url 'schooladmin:section' sectionid=section.id %}">Clear</a>
-                {% endif %}
-            </form>
-        {% endif %}
-        <div class="data-table">
-            {% render_table secitem_table 'django_tables2/bootstrap.html' %}
-        </div>
-    </div>
-    
-        <h3>Students:</h3>
-    <div id="Students">
-        {% if secstud_filter %}
-            <form id="secstud_filter" action="" method="get" class="form form-inline" autocomplete="off">
-                {% bootstrap_form secstud_filter.form layout='inline' %}
-                <button class="btn btn-primary">Filter</button>
-                {% if secstud_has_filter %}
-                    <a  class="btn btn-success" href="{%  url 'schooladmin:section' sectionid=section.id %}">Clear</a>
-                {% endif %}
-            </form>
-        {% endif %}
-        <div class="data-table-alt">
-            {% render_table secstud_table 'django_tables2/bootstrap.html' %}
-        </div>
-    </div>
+    <h3>Reference Items:</h3>
+    {% sis_filter 'secitems' %}
+
+    <h3>Students:</h3>
+    {% sis_filter 'secstuds' %}
     
   <form method="post">
     {% csrf_token %}
@@ -95,7 +69,6 @@ addButtonClickHandler('new_section', '{% url 'schooladmin:section_new_from_secti
 addButtonClickHandler('edit_section', '{% url 'schooladmin:section_edit' sectionid=section.id %}');
 addButtonClickHandler('refresh_items', '{% url 'schooladmin:section_refreshitems' sectionid=section.id %}');
 addButtonClickHandler('manage_students', '{% url 'schooladmin:section_students_manage' sectionid=section.id %}');
-addClickHandlers("{% url 'schooladmin:sectionstudent' id=1234599 %}", "sectionstudent_row");
 stylizeRange({field:'index',min_label:'From #',max_label:'To #',prefix:"secitem"});
 </script>
 {% endblock %}
