+++ student-info-system/schooladmin/templates/schooladmin/sections.html
@@ -1,30 +1,16 @@
 {% extends 'sis/subpage_base.html' %}
 {% load static %}
-{% load render_table from django_tables2 %}
+{% load sis-filter %}
 {% load bootstrap3 %}
 
 {% block content %}
 <div>
     <button class="btn btn-primary" id="section_new">Add Section to Class</button>
-    <div id="Sections">
-            {% if sections_filter %}
-                <form id="sections_filter" action="" method="get" class="form form-inline" autocomplete="off">
-                    {% bootstrap_form sections_filter.form layout='inline' %}
-                    <button class="btn btn-primary">Filter</button>
-                    {% if sections_has_filter %}
-                        <a  class="btn btn-success" href="{%  url 'schooladmin:sections' %}">Clear</a>
-                    {% endif %}
-                </form>
-            {% endif %}
-            <div class="data-table-alt scrollify-me">
-                {% render_table sections_table 'django_tables2/bootstrap.html' %}
-            </div>
-    </div>
+    {% sis_filter 'sections' %}
 </div>
 <script>
 const handler = insertScrollableDivs(applyHandlers=true);
 handler();
 addButtonClickHandler('section_new', "{% url 'schooladmin:section_new' %}");
-addClickHandlers("{% url 'schooladmin:section' sectionid=1234599 %}", "section_row");
 </script>
 {% endblock content %}
