+++ student-info-system/schooladmin/templates/schooladmin/semester.html
@@ -1,7 +1,7 @@
 {% extends 'sis/subpage_base.html' %}
 {% load static %}
 {% load bootstrap3 %}
-{% load render_table from django_tables2 %}
+{% load sis-filter %}
 
 {% block content %}
 <dl class="dl-horizontal">
@@ -22,52 +22,13 @@
 </dl>
 
     <h3>Sections:</h3>
-    <div id="Sections">
-        {% if sections_filter %}
-            <form id="sections_filter" action="" method="get" class="form form-inline" autocomplete="off">
-                {% bootstrap_form sections_filter.form layout='inline' %}
-                <button class="btn btn-primary">Filter</button>
-                {% if sections_has_filter %}
-                    <a  class="btn btn-success" href="{%  url 'schooladmin:semester' semester_id=semester.id %}">Clear</a>
-                {% endif %}
-            </form>
-        {% endif %}
-        <div class="data-table-alt">
-            {% render_table sections_table 'django_tables2/bootstrap.html' %}
-        </div>
-    </div>
+    {% sis_filter 'sections' %}
 
          <h3>Enrolled Students:</h3>
-    <div id="students">
-        {% if students_filter %}
-            <form id="students_filter" action="" method="get" class="form form-inline" autocomplete="off">
-                {% bootstrap_form students_filter.form layout='inline' %}
-                <button class="btn btn-primary">Filter</button>
-                {% if students_has_filter %}
-                    <a  class="btn btn-success" href="{%  url 'schooladmin:semester' semester_id=semester.id %}">Clear</a>
-                {% endif %}
-            </form>
-        {% endif %}
-        <div class="data-table-alt">
-            {% render_table students_table 'django_tables2/bootstrap.html' %}
-        </div>
-    </div>
+    {% sis_filter 'students' %}
 
         <h3>Professors Teaching:</h3>
-    <div id="professors">
-        {% if professors_filter %}
-            <form id="professors_filter" action="" method="get" class="form form-inline" autocomplete="off">
-                {% bootstrap_form professors_filter.form layout='inline' %}
-                <button class="btn btn-primary">Filter</button>
-                {% if professors_has_filter %}
-                    <a  class="btn btn-success" href="{%  url 'schooladmin:semester' semester_id=semester.id %}">Clear</a>
-                {% endif %}
-            </form>
-        {% endif %}
-        <div class="data-table-alt">
-            {% render_table professors_table 'django_tables2/bootstrap.html' %}
-        </div>
-    </div>
+    {% sis_filter 'professors' %}
 
      <input id="section_new" name="section_new" action="action"
            class="btn btn-info"
@@ -88,7 +49,5 @@
 <script>
 addButtonClickHandler('section_new', "{% url 'schooladmin:semester_section_new' semester.id %}");
 addButtonClickHandler('edit_semester', "{% url 'schooladmin:semester_edit' semester.id %}");
-addClickHandlers("{% url 'schooladmin:user' userid=1234599 %}", "user_row");
-addClickHandlers("{% url 'schooladmin:section' sectionid=1234599 %}", "section_row");
 </script>
 {% endblock %}
