+++ student-info-system/schooladmin/templates/schooladmin/students.html
@@ -1,29 +1,15 @@
 {% extends 'sis/subpage_base.html' %}
 {% load static %}
-{% load render_table from django_tables2 %}
+{% load sis-filter %}
 {% load bootstrap3 %}
 
 {% block content %}
 <div>
   <button class="btn btn-primary" onclick="location.href='{% url 'schooladmin:user_new' %}';">New User</button>
-    <div id="Students">
-        {% if students_filter %}
-            <form id="students_filter" action="" method="get" class="form form-inline" autocomplete="off">
-                {% bootstrap_form students_filter.form layout='inline' %}
-                <button class="btn btn-primary">Filter</button>
-                {% if students_has_filter %}
-                    <a  class="btn btn-success" href="{%  url 'schooladmin:users' %}">Clear</a>
-                {% endif %}
-            </form>
-        {% endif %}
-        <div class="data-table-alt scrollify-me">
-            {% render_table students_table 'django_tables2/bootstrap.html' %}
-        </div>
-    </div>
+    {% sis_filter 'students' %}
 </div>
   <script>
     const handler = insertScrollableDivs(applyHandlers=true);
     handler();
-    addClickHandlers("{% url 'schooladmin:user' userid=1234599 %}", "user_row");
   </script>
 {% endblock content %}
