+++ student-info-system/schooladmin/views.py
@@ -11,18 +11,23 @@ from django.utils.html import format_html
 from django.urls import reverse
 
 from sis.authentication_helpers import role_login_required
+
 from sis.models import (Course, CoursePrerequisite, Major, Professor, Section, Semester, Student,
                         SectionStudent, Profile, Message)
-from sis.utils import filtered_table
 
-from .filters import (MajorFilter, SectionFilter, SemesterFilter, UserFilter, StudentFilter,
-                      ItemFilter)
+from sis.utils import filtered_table, filtered_table2, DUMMY_ID
+
 
 from sis.filters.course import CourseFilter
+from sis.filters.major import MajorFilter
 from sis.filters.message import (FullSentMessageFilter, FullReceivedMessageFilter,
                                  SentMessageFilter, ReceivedMessageFilter)
+from sis.filters.referenceitem import ItemFilter
+from sis.filters.section import SectionFilter
 from sis.filters.sectionreferenceitem import SectionItemFilter
 from sis.filters.sectionstudent import SectionStudentFilter
+from sis.filters.semester import SemesterFilter
+from sis.filters.user import StudentFilter, UserFilter
 
 from .forms import (
     CourseCreationForm,
@@ -52,24 +57,6 @@ from sis.tables.semesters import SemestersSummaryTable, SemestersTable
 from sis.tables.users import UsersTable, FullUsersTable, StudentsTable, StudentInMajorTable
 
 
-def filtered_table2(name=None,
-                   qs=None,
-                   filter=None,
-                   table=None,
-                   request=None,
-                   page_size=25,
-                   wrap_list=True):
-    filter = filter(request.GET, queryset=qs, prefix=name)
-    # weird "{name}" thing is because the HTML field has the prefix but the Filter does
-    # NOT have it in the field names
-    has_filter = any(f'{name}-{field}' in request.GET for field in set(filter.get_fields()))
-    table_source = filter.qs
-    if wrap_list:
-        table_source = list(table_source)
-    tab = table(table_source, prefix=name + "-")
-    RequestConfig(request, paginate={"per_page": page_size, "page": 1}).configure(tab)
-    return { name: { 'name': name, 'table': tab, 'filter': filter, 'has_filter': has_filter} }
-
 @role_login_required(Profile.ACCESS_ADMIN)
 def index(request):
     return render(request, 'schooladmin/home_admin.html', request.user.profile.unread_messages())
@@ -82,12 +69,15 @@ def index(request):
 def users(request):
     return render(
         request, 'schooladmin/users.html',
-        filtered_table(
+        filtered_table2(
             name='users',
             qs=User.annotated(),
             filter=UserFilter,
             table=FullUsersTable,
             request=request,
+            scrollable=True,
+            self_url=reverse('schooladmin:users'),
+            click_url=reverse('schooladmin:user', args=[DUMMY_ID]),
         ))
 
 
@@ -95,12 +85,15 @@ def users(request):
 def students(request):
     return render(
         request, 'schooladmin/students.html',
-        filtered_table(
+        filtered_table2(
             name='students',
             qs=User.annotated().filter(profile__role=Profile.ACCESS_STUDENT),
             filter=StudentFilter,
             table=StudentsTable,
             request=request,
+            scrollable=True,
+            self_url=reverse('schooladmin:students'),
+            click_url=reverse('schooladmin:student', args=[DUMMY_ID]),
         ))
 
 
@@ -132,22 +125,24 @@ def user(request, userid):
         'user': the_user,
     }
     data.update(
-        filtered_table(
+        filtered_table2(
             name='received',
             qs=the_user.profile.sent_to.all(),
             filter=FullReceivedMessageFilter,
             table=MessageReceivedTable,
             request=request,
             wrap_list=False,
+            self_url=reverse('schooladmin:user', args=[userid]),
         ))
     data.update(
-        filtered_table(
+        filtered_table2(
             name='sent',
             qs=the_user.profile.sent_by.all(),
             filter=FullSentMessageFilter,
             table=MessageSentTable,
             request=request,
             wrap_list=False,
+            self_url=reverse('schooladmin:user', args=[userid]),
         ))
 
     return render(request, 'schooladmin/user.html', data)
@@ -176,57 +171,67 @@ def student(request, userid):
         'user': the_user,
     }
     data.update(
-        filtered_table(
+        filtered_table2(
             name='semesters',
             qs=the_user.profile.student.semesters.all(),
             filter=SemesterFilter,
             table=SemestersSummaryTable,
             request=request,
             wrap_list=False,
+            self_url=reverse('schooladmin:student', args=[userid]),
+            click_url=reverse('schooladmin:semester', args=[DUMMY_ID]),
         ))
 
     data.update(
-        filtered_table(
+        filtered_table2(
             name='history',
             qs=the_user.profile.student.course_history(),
             filter=SectionStudentFilter,
             table=StudentHistoryTable,
             request=request,
+            self_url=reverse('schooladmin:student', args=[userid]),
+            click_url=reverse('schooladmin:sectionstudent', args=[DUMMY_ID]),
         ))
 
     data.update(
-        filtered_table(
+        filtered_table2(
             name='remaining',
             qs=the_user.profile.student.remaining_required_courses(),
             filter=CourseFilter,
             table=CoursesTable,
             request=request,
+            self_url=reverse('schooladmin:student', args=[userid]),
+            click_url=reverse('schooladmin:course', args=[DUMMY_ID]),
         ))
     data.update(
-        filtered_table(
+        filtered_table2(
             name='majorcourses',
             qs=the_user.profile.student.requirements_met_list(),
             filter=CourseFilter,
             table=MajorCoursesMetTable,
             request=request,
+            self_url=reverse('schooladmin:student', args=[userid]),
+            click_url=reverse('schooladmin:course', args=[DUMMY_ID]),
         ))
     data.update(
-        filtered_table(
+        filtered_table2(
             name='received',
             qs=the_user.profile.sent_to.filter(time_archived__isnull=True),
             filter=ReceivedMessageFilter,
             table=MessageReceivedTable,
             request=request,
             wrap_list=False,
+            self_url=reverse('schooladmin:student', args=[userid]),
         ))
     data.update(
-        filtered_table(
+        filtered_table2(
             name='sent',
             qs=the_user.profile.sent_by.filter(time_archived__isnull=True),
             filter=SentMessageFilter,
             table=MessageSentTable,
             request=request,
             wrap_list=False,
+            self_url=reverse('schooladmin:student', args=[userid]),
         ))
 
     return render(request, 'schooladmin/student.html', data)
@@ -255,48 +260,56 @@ def professor(request, userid):
         'user': the_user,
     }
     data.update(
-        filtered_table(
+        filtered_table2(
             name='semesters',
             qs=the_user.profile.professor.semesters_teaching(),
             filter=SemesterFilter,
             table=SemestersSummaryTable,
             request=request,
+            self_url=reverse('schooladmin:professor', args=[userid]),
+            click_url=reverse('schooladmin:semester', args=[DUMMY_ID]),
         ))
 
     data.update(
-        filtered_table(
+        filtered_table2(
             name='sections',
             qs=the_user.profile.professor.section_set,
             filter=SectionFilter,
             table=SectionsTable,
             request=request,
+            self_url=reverse('schooladmin:professor', args=[userid]),
+            click_url=reverse('schooladmin:section', args=[DUMMY_ID]),
         ))
 
     data.update(
-        filtered_table(
+        filtered_table2(
             name='items',
             qs=the_user.profile.professor.referenceitem_set,
             filter=ItemFilter,
             table=ProfReferenceItemsTable,
             request=request,
+            self_url=reverse('schooladmin:professor', args=[userid]),
+            click_url=reverse('schooladmin:professor_item', args=[userid, DUMMY_ID]),
         ))
     data.update(
-        filtered_table(
+        filtered_table2(
             name='received',
             qs=the_user.profile.sent_to.filter(time_archived__isnull=True),
             filter=ReceivedMessageFilter,
             table=MessageReceivedTable,
             request=request,
             wrap_list=False,
+            self_url=reverse('schooladmin:professor', args=[userid]),
         ))
     data.update(
-        filtered_table(
+        filtered_table2(
             name='sent',
             qs=the_user.profile.sent_by.filter(time_archived__isnull=True),
             filter=SentMessageFilter,
             table=MessageSentTable,
             request=request,
             wrap_list=False,
+            self_url=reverse('schooladmin:professor', args=[userid]),
         ))
 
     return render(request, 'schooladmin/professor.html', data)
@@ -316,12 +329,14 @@ def professor_items(request, userid):
         'user': the_user,
     }
     data.update(
-        filtered_table(
+        filtered_table2(
             name='items',
             qs=the_user.profile.professor.referenceitem_set,
             filter=ItemFilter,
             table=ProfReferenceItemsTable,
             request=request,
+            self_url=reverse('schooladmin:professor_items', args=[userid]),
+            click_url=reverse('schooladmin:professor_item', args=[userid, DUMMY_ID]),
         ))
 
     return render(request, 'schooladmin/professor_items.html', data)
@@ -545,12 +560,15 @@ def user_new(request):
 def majors(request):
     return render(
         request, 'schooladmin/majors.html',
-        filtered_table(
+        filtered_table2(
             name='majors',
             qs=Major.objects,
             filter=MajorFilter,
             table=MajorsTable,
             request=request,
+            scrollable=True,
+            self_url=reverse('schooladmin:majors'),
+            click_url=reverse('schooladmin:major', args=[DUMMY_ID]),
         ))
 
 
@@ -569,40 +587,48 @@ def major(request, majorid):
         'permit_edit': request.user.profile.role == Profile.ACCESS_ADMIN,
     }
     data.update(
-        filtered_table(
+        filtered_table2(
             name='profs',
             qs=User.annotated().filter(profile__professor__major=the_major),
             filter=UserFilter,
             table=UsersTable,
             request=request,
+            self_url=reverse('schooladmin:major', args=[majorid]),
+            click_url=reverse('schooladmin:professor', args=[DUMMY_ID]),
         ))
 
     data.update(
-        filtered_table(
+        filtered_table2(
             name='required',
             qs=Course.objects.filter(required_by=the_major),
             filter=CourseFilter,
             table=CoursesTable,
             request=request,
+            self_url=reverse('schooladmin:major', args=[majorid]),
+            click_url=reverse('schooladmin:course', args=[DUMMY_ID]),
         ))
 
     data.update(
-        filtered_table(
+        filtered_table2(
             name='offered',
             qs=Course.objects.filter(major=the_major),
             filter=CourseFilter,
             table=CoursesForMajorTable,
             request=request,
+            self_url=reverse('schooladmin:major', args=[majorid]),
+            click_url=reverse('schooladmin:course', args=[DUMMY_ID]),
         ))
 
     if include_students:
         data.update(
-            filtered_table(
+            filtered_table2(
                 name='students',
                 qs=User.annotated().filter(profile__student__major=the_major),
                 filter=UserFilter,
                 table=StudentInMajorTable,
                 request=request,
+                self_url=reverse('schooladmin:major', args=[majorid]),
+                click_url=reverse('schooladmin:student', args=[DUMMY_ID]),
             ))
 
     return render(request, 'schooladmin/major.html', data)
@@ -651,12 +677,15 @@ def major_new(request):
 def courses(request):
     return render(
         request, 'schooladmin/courses.html',
-        filtered_table(
+        filtered_table2(
             name='courses',
             qs=Course.objects.all(),
             filter=CourseFilter,
             table=CoursesTable,
             request=request,
+            scrollable=True,
+            self_url=reverse('schooladmin:courses'),
+            click_url=reverse('schooladmin:course', args=[DUMMY_ID]),
         ))
 
 
@@ -671,30 +700,36 @@ def course(request, courseid):
         'course': the_course,
     }
     data.update(
-        filtered_table(
+        filtered_table2(
             name='sections',
             qs=Section.objects.filter(course=the_course),
             filter=SectionFilter,
             table=SectionForClassTable,
             request=request,
+            self_url=reverse('schooladmin:course', args=[courseid]),
+            click_url=reverse('schooladmin:section', args=[DUMMY_ID]),
         ))
     data.update(
-        filtered_table(
+        filtered_table2(
             name='prereqs',
             qs=Course.objects.filter(a_prerequisite__course=the_course),
             filter=CourseFilter,
             table=CoursesTable,
             request=request,
             page_size=5,
+            self_url=reverse('schooladmin:course', args=[courseid]),
+            click_url=reverse('schooladmin:course', args=[DUMMY_ID]),
         ))
     data.update(
-        filtered_table(
+        filtered_table2(
             name='neededby',
             qs=Course.objects.filter(a_course__prerequisite=the_course),
             filter=CourseFilter,
             table=CoursesTable,
             request=request,
             page_size=10,
+            self_url=reverse('schooladmin:course', args=[courseid]),
+            click_url=reverse('schooladmin:course', args=[DUMMY_ID]),
         ))
 
     return render(request, 'schooladmin/course.html', data)
@@ -750,6 +785,9 @@ def semesters(request):
             filter=SemesterFilter,
             table=SemestersTable,
             request=request,
+            scrollable=True,
+            self_url=reverse('schooladmin:semesters'),
+            click_url=reverse('schooladmin:semester', args=[DUMMY_ID]),
         ))
 
 
@@ -764,28 +802,34 @@ def semester(request, semester_id):
         'semester': the_semester,
     }
     data.update(
-        filtered_table(
+        filtered_table2(
             name='sections',
             qs=Section.objects.filter(semester=the_semester),
             filter=SectionFilter,
             table=SectionsTable,
             request=request,
+            self_url=reverse('schooladmin:semester', args=[semester_id]),
+            click_url=reverse('schooladmin:section', args=[DUMMY_ID]),
         ))
     data.update(
-        filtered_table(
+        filtered_table2(
             name='students',
             qs=the_semester.students_attending(),
             filter=StudentFilter,
             table=StudentsTable,
             request=request,
+            self_url=reverse('schooladmin:semester', args=[semester_id]),
+            click_url=reverse('schooladmin:student', args=[DUMMY_ID]),
         ))
     data.update(
-        filtered_table(
+        filtered_table2(
             name='professors',
             qs=the_semester.professors_teaching(),
             filter=UserFilter,
             table=UsersTable,
             request=request,
+            self_url=reverse('schooladmin:semester', args=[semester_id]),
+            click_url=reverse('schooladmin:professor', args=[DUMMY_ID]),
         ))
 
     return render(request, 'schooladmin/semester.html', data)
@@ -844,12 +888,15 @@ def semester_section_new(request, semester_id):
 def sections(request):
     return render(
         request, 'schooladmin/sections.html',
-        filtered_table(
+        filtered_table2(
             name='sections',
             qs=Section.objects.all(),
             filter=SectionFilter,
             table=SectionsTable,
             request=request,
+            scrollable=True,
+            self_url=reverse('schooladmin:sections'),
+            click_url=reverse('schooladmin:section', args=[DUMMY_ID]),
         ))
 
 
@@ -864,20 +911,23 @@ def section(request, sectionid):
         'section': the_section,
     }
     data.update(
-        filtered_table(
-            name='secstud',
+        filtered_table2(
+            name='secstuds',
             qs=SectionStudent.objects.filter(section=the_section),
             filter=SectionStudentFilter,
             table=StudentInSectionTable,
             request=request,
+            self_url=reverse('schooladmin:section', args=[sectionid]),
+            click_url=reverse('schooladmin:sectionstudent', args=[DUMMY_ID]),
         ))
     data.update(
-        filtered_table(
-            name='secitem',
+        filtered_table2(
+            name='secitems',
             qs=the_section.sectionreferenceitem_set,
             filter=SectionItemFilter,
             table=ReferenceItemsForSectionTable,
             request=request,
+            self_url=reverse('schooladmin:section', args=[sectionid]),
         ))
 
     return render(request, 'schooladmin/section.html', data)
@@ -1070,22 +1120,24 @@ def profile(request):
         'user': the_user,
     }
     data.update(
-        filtered_table(
+        filtered_table2(
             name='received',
             qs=the_user.profile.sent_to.all(),
             filter=FullReceivedMessageFilter,
             table=MessageReceivedTable,
             request=request,
             wrap_list=False,
+            self_url=reverse('schooladmin:profile'),
         ))
     data.update(
-        filtered_table(
+        filtered_table2(
             name='sent',
             qs=the_user.profile.sent_by.all(),
             filter=FullSentMessageFilter,
             table=MessageSentTable,
             request=request,
             wrap_list=False,
+            self_url=reverse('schooladmin:profile'),
         ))
 
     return render(request, 'schooladmin/profile.html', data)
