+++ student-info-system/sis/tables/__init__.py
@@ -1,6 +1,14 @@
 import django_tables2 as tables
 
 
+# mix in a method to get specified CSS row class
+def row_class(cls):
+    ra = getattr(cls, 'Meta').row_attrs
+    return ra['class']
+
+tables.Table.row_class = classmethod(row_class)
+
+
 # Each column and cell has its own CSS class based on the type of
 # data in it.
 def field_css_classes(field_name):
