--- student-info-system/sis/templates/sis/sis-filter-tag.html
@@ -1,12 +1,0 @@
-{% load static %}
-{% load bootstrap3 %}
-
-{% if filt %}  
-  <form id="{{ filt }}" action="" method="get" class="form form-inline" autocomplete="off">
-    {% bootstrap_form filt.form layout='inline' %}
-    <button class="btn btn-primary">Filter</button>
-    {% if has_filter %}
-      <a class="btn btn-success" href="{% url destination %}">Clear</a>
-    {% endif %}
-  </form>
-{% endif %}
