+++ student-info-system/sis/templatetags/sis-filter.py
@@ -5,13 +5,9 @@ register = template.Library()
 
 
 @register.inclusion_tag('sis/sis-filter.html',takes_context=True)
-def sis_filter(context, filt, destination):
-    """Uses django-filter and bootstrap. Requires a filter "group name", and
-       the destination url for the page"""
-
+def sis_filter(context, filt):
     context = {
         'filt' : context[filt],
-        'destination': destination,
         'request': context['request'],
     }
     return context
