+++ student-info-system/sis/utils.py
@@ -1,5 +1,7 @@
 from django_tables2 import RequestConfig
 
+# used as dummy in addClickHandler
+DUMMY_ID = 1234599
 
 # helper function to make tables
 # merge the result of this into the response data
@@ -25,3 +27,35 @@ def filtered_table(name=None,
     tab = table(table_source, prefix=name + "-")
     RequestConfig(request, paginate={"per_page": page_size, "page": 1}).configure(tab)
     return {name + '_table': tab, name + '_filter': filt, name + '_has_filter': has_filter}
+
+def filtered_table2(name=None,
+                   qs=None,
+                   filter=None,
+                   table=None,
+                   request=None,
+                   page_size=25,
+                   row_class=None,
+                   self_url=None,
+                   click_url=None,
+                   scrollable=False,
+                   table_type='data-table-alt',
+                   wrap_list=True):
+    if row_class is None:
+        # "table" is the class, for which we added...
+        row_class = table.row_class()
+    filter = filter(request.GET, queryset=qs, prefix=name)
+    # weird "{name}" thing is because the HTML field has the prefix but the Filter does
+    # NOT have it in the field names
+    has_filter = any(f'{name}-{field}' in request.GET for field in set(filter.get_fields()))
+    table_source = filter.qs
+    if wrap_list:
+        table_source = list(table_source)
+    tab = table(table_source, prefix=name + "-")
+    div_classes = table_type
+    if scrollable:
+        div_classes += ' scrollify-me'
+    RequestConfig(request, paginate={"per_page": page_size, "page": 1}).configure(tab)
+    return { name: { 'name': name, 'table': tab,
+                     'filter': filter, 'has_filter': has_filter,
+                     'self_url': self_url, 'click_url': click_url,
+                     'row_class': row_class, 'div_classes':div_classes, } }
