+++ student-info-system/student/templates/student/history.html
@@ -1,7 +1,7 @@
 {% extends 'sis/subpage_base.html' %}
 {% load static %}
 {% load bootstrap3 %}
-{% load render_table from django_tables2 %}
+{% load sis-filter %}
 {%  load user_demographics %}
 
 {% block content %}
@@ -14,20 +14,7 @@
   <dd>{{ user.profile.student.credits_earned }}</dd>
 </dl>
     <h3>Course History:</h3>
-    <div id="history">
-        {% if history_filter %}
-            <form id="history_filter" action="" method="get" class="form form-inline" autocomplete="off">
-                {% bootstrap_form history_filter.form layout='inline' %}
-                <button class="btn btn-primary">Filter</button>
-                {% if history_has_filter %}
-                    <a  class="btn btn-success" href="{%  url 'student:history' %}">Clear</a>
-                {% endif %}
-            </form>
-        {% endif %}
-        <div class="data-table-alt">
-            {% render_table history_table 'django_tables2/bootstrap.html' %}
-        </div>
-    </div>
+    {% sis_filter 'history' %}
 
  <form method="post">
     {% csrf_token %}
@@ -41,6 +28,5 @@
 
   </form>
 <script>
-addClickHandlers("{% url 'student:sectionstudent' id=1234599 %}", "sectionstudent_row");
 </script>
 {% endblock %}
