+++ student-info-system/student/templates/student/items.html
@@ -1,7 +1,7 @@
 {% extends 'sis/subpage_base.html' %}
 {% load static %}
 {% load bootstrap3 %}
-{% load render_table from django_tables2 %}
+{% load sis-filter %}
 
 {% block content %}
 <dl class="dl-horizontal">
@@ -10,20 +10,7 @@
     <dd><a href="{% url 'schooladmin:semester' semester_id=semester.id %}">{{ semester }}</a></dd>
 </dl>
 
-<div id="items">
-    {% if secitem_filter %}
-        <form id="secitem_filter" action="" method="get" class="form form-inline" autocomplete="off">
-            {% bootstrap_form secitem_filter.form layout='inline' %}
-            <button class="btn btn-primary">Filter</button>
-            {% if secitem_has_filter %}
-                <a  class="btn btn-success" href="{%  url 'student:secitems' %}">Clear</a>
-            {% endif %}
-        </form>
-    {% endif %}
-    <div class="data-table">
-        {% render_table secitem_table 'django_tables2/bootstrap.html' %}
-    </div>
-</div>
+{% sis_filter 'items' %}
 <div>
     <input
         action="action"
@@ -33,7 +20,6 @@
     />
 
 <script>
-addClickHandlers("{% url 'student:secitem' id=1234599 %}", "secitem_row");
 stylizeRange({field:'index',min_label:'From #',max_label:'To #',prefix:"secitem"});
 </script>
 {% endblock %}
